Bug 31383: Fix the DB upgrade process
This patch fixes misc4dev/run_tests.pl --run-db-upgrade-only

We need to make sure the column exists before running the query.

+++ installer/data/mysql/db_revs/211200005.pl
@@ -6,15 +6,18 @@ return {
     up => sub {
         my ($args) = @_;
         my ($dbh, $out) = @$args{qw(dbh out)};
-        $dbh->do(q{
-            DELETE FROM additional_contents
-            WHERE code NOT IN (
-                SELECT code FROM (
-                    SELECT code
-                    FROM additional_contents
-                    WHERE lang = "default"
-                ) as tmp
-            );
-        });
+        # Run this only if 230600041 has not been run before
+        if ( column_exists('additional_contents', 'lang') ) {
+            $dbh->do(q{
+                DELETE FROM additional_contents
+                WHERE code NOT IN (
+                    SELECT code FROM (
+                        SELECT code
+                        FROM additional_contents
+                        WHERE lang = "default"
+                    ) as tmp
+                );
+            });
+        }
     },
 }
